feat: 适配 F1 26 (2026 Season Pack) UDP 遥测格式
- 版本化解析：UdpPacketConstants 增加 2025/2026 双版本尺寸表与 GetMaxCarsInSession
- FixedSizePacketParser 支持按 packetFormat 校验长度，解析器内按版本分支
- 适配 26 布局：22→24 车、Motion G 力 int16 量化、engineTemperature 缩窄、
  Participants ID 加宽 uint16、CarStatus 新增 ErsHarvestedLimitPerLap、
  Session 新增 active aero/DRS 区段与辅助字段、Event Collision severity
- 新增 CarTelemetry2 包 (PacketId=16, 269B) 及解析器
- PacketDispatcher 增加未知格式校验，仅支持 2025/2026
- 下游：CarSnapshot.TeamId ushort 化、CarStateStore 硬编码 22→24 上限
- RawLogAnalyzer 支持 CarTelemetry2
- 新增 PacketParserFormat2026Tests 25 用例

+++ F1Telemetry.Core/Models/CarSnapshot.cs
@@ -43,7 +43,7 @@ public sealed record CarSnapshot
     /// <summary>
     /// Gets the raw team identifier from the UDP packet.
     /// </summary>
-    public byte? TeamId { get; init; }
+    public ushort? TeamId { get; init; }
 
     /// <summary>
     /// Gets the raw nationality identifier from the UDP packet.
